jpa delete, detail

+++ src/main/java/kr/co/lir/service/BoardService.java
@@ -9,6 +9,6 @@ public interface BoardService {
 
     List<Board> boardList();
     void boardInsert(Board board);
-    void boardDelete(Integer id);
-    Board boardDetail(Integer id);
+    void boardDelete(Long id);
+    Board boardDetail(Long id);
 }
